projects/content-automation-studio/backend/requirements.txt
Flask==2.3.0
Flask-CORS==4.0.0
webdavclient3==3.14.6
requests==2.31.0
google-cloud-aiplatform==1.35.0
google-generativeai==0.3.2
Pillow==10.1.0
python-dotenv==1.0.0
gunicorn==21.2.0